backend/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
requests
google-genai